+++ dart/geobase/lib/src/vector_data/model/feature/feature_collection.dart
@@ -9,6 +9,8 @@ import 'package:meta/meta.dart';
 import '/src/coordinates/base.dart';
 import '/src/utils/property_builder.dart';
 import '/src/vector/content.dart';
+import '/src/vector/encoding.dart';
+import '/src/vector/formats.dart';
 import '/src/vector_data/model/bounded.dart';
 import '/src/vector_data/model/geometry.dart';
 
@@ -95,7 +97,45 @@ class FeatureCollection<E extends Feature> extends Bounded {
   /// property data outside it is stored in this member.
   Map<String, Object?>? get custom => _custom;
 
-  // todo: toString
+  /// Writes this feature collection to [writer].
+  void writeTo(FeatureContent writer) {
+    final cust = custom;
+    writer.featureCollection(
+      (feat) {
+        for (final item in features) {
+          item.writeTo(feat);
+        }
+      },
+      count: features.length,
+      custom: cust != null
+          ? (props) {
+              cust.forEach((name, value) {
+                props.property(name, value);
+              });
+            }
+          : null,
+    );
+  }
+
+  /// The string representation of this feature collection, with [format]
+  /// applied.
+  ///
+  /// When [format] is not given, then [GeoJSON] is used as a default.
+  ///
+  /// Use [decimals] to set a number of decimals (not applied if no decimals).
+  String toStringAs({
+    TextWriterFormat<FeatureContent> format = GeoJSON.feature,
+    int? decimals,
+  }) {
+    final encoder = format.encoder(decimals: decimals);
+    writeTo(encoder.writer);
+    return encoder.toText();
+  }
+
+  /// The string representation of this feature collection as specified by
+  /// [GeoJSON].
+  @override
+  String toString() => toStringAs();
 
   @override
   bool operator ==(Object other) =>